# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/request_ship/{id}/ships/{ship_id}": {
            "put": {
                "description": "Update the number of ships in a specific request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "request_ships"
                ],
                "summary": "Update ship count in request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Ship ID",
                        "name": "ship_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated ship count",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "ships_count": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status: string, message: string",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "description: string",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "error: string",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a ship from a specific request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "request_ships"
                ],
                "summary": "Delete ship from request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Ship ID",
                        "name": "ship_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "description: string",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "status: string, description: string",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "status: string, description: string",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/api/ships": {
            "get": {
                "description": "Retrieve a list of ships with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ships"
                ],
                "summary": "Get list of ships",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ship name filter",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Minimum capacity filter",
                        "name": "capacity",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Active status filter",
                        "name": "is_active",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data: []ds.Ship, count: int",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "error: string",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new ship to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ships"
                ],
                "summary": "Create a new ship",
                "parameters": [
                    {
                        "description": "Ship data",
                        "name": "ship",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Ship"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "data: ds.Ship",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "error: string",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "error: string",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/ships/{id}": {
            "get": {
                "description": "Retrieve details of a specific ship by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ships"
                ],
                "summary": "Get a single ship",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ship ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data: ds.Ship",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "error: string",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "error: string",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "description": "Update details of an existing ship by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ships"
                ],
                "summary": "Update a ship",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ship ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated ship data",
                        "name": "ship",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Ship"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data: ds.Ship",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "error: string",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "error: string",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a ship from the system by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ships"
                ],
                "summary": "Delete a ship",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ship ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: string",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "message: string",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "error: string",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/ships/{id}/add-to-ship-bucket": {
            "post": {
                "description": "Add a ship to a user's request draft",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ships"
                ],
                "summary": "Add ship to request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ship ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: string, data: {request_ship_id: int, ship_id: int}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "message: string",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "message: string",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "status: string, description: string",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/ships/{id}/image": {
            "post": {
                "description": "Upload an image for a specific ship",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ships"
                ],
                "summary": "Upload ship image",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ship ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Image file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Image file (alternative)",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data: {ship_id: int, photo_url: string, message: string}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "message: string",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "message: string",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "message: string",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ds.Ship": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "number"
                },
                "containers": {
                    "type": "integer"
                },
                "cranes": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "draft": {
                    "type": "number"
                },
                "length": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "photoURL": {
                    "type": "string"
                },
                "shipID": {
                    "type": "integer"
                },
                "width": {
                    "type": "number"
                }
            }
        },
...
    }
}